Rename Handle to ZiplineReference
Like AtomicReference and WeakReference it creates a special ability
for its subject.

Also add diagnostic APIs to Zipline to get the list of active
services both published and consumed. This is intended to be useful
in detecting leaked objects.

Also implement close() that works on a ZiplineReference and can be
called by either inbound or outbound side.