Don't encode parameter lists as bytes
On QuickJS the performance of StringBuilder is quadratic. This impacts lots
of code in Zipline, but decoding large JSON bodies from UTF-8 is particularly
problematic.

Time to call ByteArray.concatToString() running QuickJS on my MacBook Pro

| Array Size   | Time (seconds) |
| -----------: | :------------- |
|         100  | 0              |
|       1.000  | 0              |
|      10,000  | 0              |
|     100,000  | 1              |
|   1.000,000  | 24             |
|  10,000,000  | 3600+          |

We will want to solve this in QuickJS, Kotlin, and/or Okio. For now I'm
punting on the problem by not dealing with bytes where it isn't necessary.
As a bonus, the code is simpler.