Do not use contextual serialization for ZiplineReference
This causes the serialization library to attempt to resolve a serializer for the type parameter which is almost certainly not a serializable type. This magically worked on the JVM and JS but eagerly fails on Native and thus required a different approach. Instead, we directly resolve the serializer for this type which is now held on the Endpoint and completely ignores the generic type parameter.

Also move EndpointTest and ZiplineReferenceTest into engineTest/ to get them running on native.