ZiplineService
This introduces ZiplineService (public API) and ZiplineServiceAdapter (internal)
that should soon replace other classes including InboundBridge, OutboundBridge,
and ZiplineReference.

The biggest benefit of this change is that it should make it possible to not
need the clumsy ZiplineReference box when passing interfaces across the bridge.

A side-effect of this change is that API-dependent code is now generated
alongside the interface and not the use-sites. This should fix our problems
with incremental builds. (Without this, any time an interface method is changed
but the calls to Zipline.get() or Zipline.set() don't change, a rebuild is
not triggered and things fail at runtime.)

https://github.com/cashapp/zipline/issues/383

In follow-up PRs I'd like to tear down the above features obsoleted by this
change. I'm holding off on doing that here to simplify the transition. As a
consequence, this PR introduces some duplicate code.