Report function signature on unexpected function
When faced with an unexpected function (e.g., via incompatible APIs)
we currently just display the function name. This isn't super
useful if the function name is the same but the parameter types or
return type differs.

Through this change, we now display the whole function signature
(except for the function location due to our use of duck typing),
so instead of `echo` we'd have
`fun echo(com.example.Request): com.example.Response`.

This is the first part of the solution to #402.