Support serializing ZiplineServices
Currently this only works if the service is a parameter or return
type of another ZiplineService.

I tried quite hard to make it work everywhere using @Serializable
but that required some pretty ugly compromises. In particular we'd
need to either register services explicitly, or lean on implementation
details of kotlinx.serialization and convert ZiplineService from an
interface to an abstract class.