Make the FlowReference wrapper unnecessary
This drops support for Flows as fields in objects passed
through Zipline services. But flows can be fields in any
parameter or return type. This is the same limitation of
ZiplineService instances.

All flows passed to or returned from a ZiplineService is
exactly one-shot - not consuming or double-consuming are
both unsupported.