Introduce Zipline.json
This is the best way to encode and decode variables with
Zipline because it supports contextual serialization. Using
serializers alone is not sufficient.

This also simplifies endpoint initialization on Kotlin/JS.
The old code was mutable to work around initialization cycles.
With this we just forward calls to CallChannel to an instance
that might not be ready when the forwarder is created.