Add more events to EventListener (#468)
I haven't called out whether the host is the caller or the receiver
of the call. Is this useful?

I also haven't made suspending functions special. My assumption is
the implementation will do fancy things to make the suspend callback
look like a single regular function call.

Lastly I think we always implement this interface on the host platform
but call it from either platform. That'll require some asymmetry.