Promote some loader tests to commonTest (#599)
* Promote some loader tests to commonTest

Unfortunately there's a bug where Okio's FakeFileSystem doesn't
work on Kotlin 1.7.0 on native due to a transitive dependency on
kotlinx-datetime.

https://youtrack.jetbrains.com/issue/KT-52554

I'm going to need to update Okio and ship a release. In the interim
just use the regular file system for these tests.

* Suppress Android tests

Previously we didn't have any. Now that I've promoted tests
to commonMain we have some, but they don't run properly.