Change how throwables are serialized (#609)
This PR is backwards-incompatible and applications built with
this will not work with runtimes built on an earlier version,
or vice versa.

This splits serialized exceptions to express type information
separately. It will allow us to customize which exception types
are propagated. Initially this is used to make handing a
ZiplineApiMismatchException easy and safe.