Endpoint-layer support for shutdown.
This is mostly just book-keeping of how many outbound services
are present. When the last of them is closed then the overall
endpoint can be closed.

This treats closing as a mechanism that is local to a single
endpoint. In practice this works for our use case, as only the
host application is interested in ever shutting down the QuickJS
VM, it is not interested in shutting itself down.

Note that this keeps counts of client-side outbound interfaces
instead of references to those interfaces directly. That would
be simpler but it would interfere with leak detection as the
Endpoint would always hold a strong reference.