Unhack serializer lookups
This simplifies the internals of generated ZiplineServiceAdapter
subtypes. Instead of accepting a List<Ktype> and a List<KSerializer>
it now accepts just a List<KSerializer>, and the set of types in that
list is much simpler - one element per type parameter.

The previous definition was more complex - the types were for lazy
lookups (with a SerializersModule) and the serializers were for eager
lookups (for flows and pass-by-reference).

The next step when this is done is supporting for @Contextual and
other serialization annotations.

https://github.com/cashapp/zipline/issues/602