Make events more complete
This adds a new 'Zipline' parameter to the EventListener where
appropriate. I'm hoping we can use this to track which Zipline
instance an event came through in applications where there is
more than one.

This also adds an event when the load is skipped because that code
is unchanged. This fixes a small bug where we emitted a start
event but not an end event.