Scope EventListener to a specific Zipline instance
This introduces a new interface, EventListener.Factory, similar to
a class of the same name in OkHttp.

Implementations of this interface may produce a distinct EventListener
instance for each Zipline instance. We want this downstream so that we
can associate metadata collected at loading time to events that occur
at execution time.